Fixed wrath spectral blades not proccing combo or processing enchants properly.

+++ core/src/main/java/com/zrp200/rkpd2/actors/hero/abilities/rat_king/Wrath.java
@@ -150,7 +150,7 @@ public class Wrath extends ArmorAbility {
         });
     }
     public static class SpectralBladesTracker extends Talent.SpiritBladesTracker {
-        public int effectiveness;
+        public float effectiveness;
     }
 
     private void finish() {
